Added new classes and basic annotation process
The now implemented function to annotate objects of the class Tad
iterates through all Tads and searches for overlapping genes and
enhancers in the same chromosome. To save memory and time I deleted
every gene/enhancer that ended in the current TAD and iterated through
both genes and enhancers at the same time. The function returns a
dictionary were the keys are the chromosomes and the values lists of Tad
objects.