Initial implementation of CNV annotation
I implemented a basic version of the CNV annotation. At this time the
output is similar to the annotated TADs: a dictionary with chromosomes
as keys and lists of CNVs with their overlapping TADs. There were still
some errors in the annotation process, which is why I had to change the
tests again.