New annotation protocol and methods for filtering visualizing
The default annotation of TADs now requires genes and enhancers that are annotated with pLI,LOEUF and Phastcon score. Additionally the ID is saved. TADs can now be annotated for two basic binary features: 1 if the TAD contains a high pLI gene (>= 0.9), 1 if the TAD contains an enhancer with a high Phastcon score (>=0.6).