New Scikit pipeline with visualizations
I have implemented a pipeline which can be used to train and test differen classifiers. Currently the user has to preprocess their CNV data with the preprocessing script and the resulting pandas Dataframes are then used in classification. In addition to classification metrics and ROC curves the correlation inbetween indepenedent features and between independent and the dependent feature can be visualized with a heatmap.