Implemented a more generilzed version of Annotation
In the previous version Genes and Enhancer had to specified specifically for the annotation of TADs. Now every bed file can be used. The name of the annotations (e.g. genes) has to be given as input by the user. The input BED file needs a header marked with # if any additional columns are supposed to be used in further analysis.