Added parameter tuning, feature importance by permutation and additonal plotting
The previous feature importance calculation was based on the gini impurity of the trained model. Since this method is affected by collinear features I implemented a permutation based method. I also added a step to cluster features before permuting according to their correlation i.e. hierachical clustering. With this commit a shema for parameter tuning was added to the classifier script.